currencies_direct/requirements.txt
tensorflow==1.14.0
keras-rl==0.4.2
gym==0.16.0
h5py==2.10.0